Camera2Video: Update dynamic range profile argument from Int to Long following changes to the API.

+++ Camera2Video/app/src/main/java/com/example/android/camera2/video/fragments/DynamicRangeFragment.kt
@@ -77,9 +77,9 @@ class DynamicRangeFragment : Fragment() {
 
         private data class DynamicRangeInfo(
                 val name: String,
-                val value: Int)
+                val value: Long)
 
-        private fun dynamicRangeProfileString(value: Int) = when (value) {
+        private fun dynamicRangeProfileString(value: Long) = when (value) {
             DynamicRangeProfiles.STANDARD -> "STANDARD"
             DynamicRangeProfiles.HLG10 -> "HLG10"
             DynamicRangeProfiles.HDR10 -> "HDR10"
